Add DeoptPatcher::onPatch(), refactor init() into onLink()
Summary:
The init() method is poorly named, in that it's a function that gets called
after the patcher's link() method is called.

Similarly, this diff adds an onPatch() callback that gets run after patch() is
called.  This simplifies a bit on how DeoptPatcher subclasses clean themselves
up post-patching, and lets us do more testing in deopt_patcher_test.cpp.

Differential Revision: D58529861

fbshipit-source-id: e3562935457e458a8cd8347ffbbf140e7ffea750