Fixup jumps
Summary:
3.12 replaces `JUMP_ABSOLUTE` with `JUMP_BACKWARD` getting rid of absolute jumps and making them relative negative jumps.

CPython has some opcodes which are greater than 255 which don't ever get emitted in the final byte code. One of those is `JUMP` and it just gets transformed into a `JUMP_BACKWARD` or `JUMP_FORWARD` depending on how it's jumping. We do the same thing, replacing the existing `JUMP_ABSOLUTE` emits with a helper `emitJump` which can do an absolute or the fake `JUMP` which we'll fixup at code gen time.

The flow graph optimizer needed to be updated to support the new JUMP instructions as well, this is a lot of code duplication just to change some jumps, but ultimately we'll condense back into a single version so this duplication will be temporary.

Also fixes a small issue where we can't run single 3.12 code sbs tests because they include "." in the name.

Reviewed By: alexmalyshev

Differential Revision: D58439900

fbshipit-source-id: 95d697a2e2e39f227b595fe4297c54bf694485c1