Fix Cinder JIT crash in Python test suite
Summary:
Trace in P1416211102.

Hard to see exactly what's happening, but there's a crash in the test suite
where asmjit tries to copy a code section's data and it hits a nullptr.  This
specifically happens with the multi-section code allocator.

It seems like it's happening when one of the code sections is empty, i.e. has
never been switched to with CodeSectionOverride and had code written into it.
The chances of this being the hot section are pretty much zero; I'm pretty sure
it's the cold section that's empty.  We need to tweak the logic in addCode() to
account for empty sections.

Added an ASM_CHECK_THROW to the call that creates the cold section as well, just
in case. We're currently ignoring its `Error` return value.

Differential Revision: D58593892

fbshipit-source-id: ccef6dfc0856f1c29721788a414915e76f60f954