CinderX "compiling" against 3.12
Summary:
Attempt to get as much of CinderX compiling against 3.12 as possible by hiding the parts broken by API changes behind `#ifdef`s. The goals are to start unblocking parallel work on the codebase, and highlight areas which need upgrade work.

The focus of the work here is just to unblock compilation. Nothing should actually be upgraded, or refactored for upgrade by these changes.

The most interesting parts of this diff are:
* `Upgrade/upgrade_assert.h` - has an `UPGRADE_ASSERT()` macro which can be used in `#ifdef`s to stop execution when an unfixed area of code is entered. Also has a manually currated enum of "tags" which can be used with this macro to indicate broad category of upgrade work needed.

* `Upgrade/upgrade_stubs.[h, cpp]` - Stubbed implementation of APIs missing from 3.10.cinder when building against 3.12. Every function here compiles down to a "stub" which just assert when called.

* `Upgrade/upgrade_exports.h` - Similar to `upgrade_stubs.h` but the functions do actually exist in 3.12 but just aren't exported by `libpython3.12`.

Other notes:
* The orders of some `#include`s have been deliberately changed to avoid global name clashes with new macros.

Reviewed By: alexmalyshev

Differential Revision: D58312431

fbshipit-source-id: a0b1d5ca2ab399c9c3a944fbb0d088722ff4cea5