Introduce UPGRADE_NOTE() and use for EXPORT_JIT_OFFSETS_FOR_STROBELIGHT
Summary:
This is similar in spirit to `UPGRADE_ASSERT()` but doesn't actually assert. Instead it provides a "structured comment" we can grep for, strongly pushing the author to add a task. It could also be upgraded to an assert in future when trying to thrash out the long tail of issues.

This should be used sparingly as overuse will lead to hard to debug crashes. However, it can be useful to quickly unblock things. In this case `EXPORT_JIT_OFFSETS_FOR_STROBELIGHT` only matters when PyPerf is in use and perhaps isn't even needed for that anymore.

Reviewed By: alexmalyshev

Differential Revision: D58709161

fbshipit-source-id: 4923ba763c728f980e092efce914bc62fb029e05