Update README with license and add CONTRIBUTING
Summary:
Our README is supposed to mention our license. Added suggested verbiage.

We're also supposed to have a CONTRIBUTING.md file. This adds a rough initial version of that. Given that you can't really build CinderX outside of Meta this just roughly suggests trying to do a PR against Cinder and we'll deal with the fallout. When we get to a point where we actually have the ability to build CinderX we can have more helpful instructions.

Reviewed By: alexmalyshev

Differential Revision: D59164050

fbshipit-source-id: f70b6a0b25fe9032f5eb48c415c2bed0cb11e8db