Hackily fix ASAN crash in JIT'd code for _Py_RefTotal
Summary:
There's an issue where asmjit fails to relocate the address of &_Py_RefTotal
when built in debug mode plus ASAN instrumentation.  It consistenly happens with
this one global variable and not with others.  I've narrowed it down to the
compiler not liking our use of

  total = Move MemImm{&_Py_RefTotal}

but I can't explain exactly why it's happening.  The way to avoid this crash is
to explicitly move the address into a register and then load from that. The generated
code here is only generated for debug builds so its performance doesn't matter
very much.  It's not a great solution, but it's a useful band-aid for now.  We know we
have issues with MemImm being "weird" that we need to fix.

This also needed a change to add BranchNE to LIR, because somehow the backend is
now tweaking the BranchE instruction and trying to negate it, which wasn't
supported before.

Differential Revision: D59169443

fbshipit-source-id: 669a8730b2bc4f9bced756c6b1ab593d87ef127a