Add BranchNE to LIR
Summary:
We currently have BranchE and we use it for IncRef LIR generation.  However it
cannot be negated with Instruction::negateBranchCC(), which can cause an
assertion error if the code gets optimized differently than it does today.

This diff adds BranchNE so that we can negate BranchE.  Technically speaking,
BranchE/BranchNE are identical to BranchZ/BranchNZ in x86-64, which we already
have in LIR.  However the "E" variants make more sense when you look at how
we're creating them from Equal and NotEqual HIR comparisons.

Differential Revision: D59169441

fbshipit-source-id: 61b10acfdea4be4e60bea9eba276eb276bf17719