Replace Ci_cinderx_initialized with CinderX module variable
Summary:
In 3.10 we had a hook to determine if the CinderX module had been initialized. This was most easily implemented as a global native variable as it was checked in the Cinder/CPython runtime. With CinderX hopefully we won't need need any checks in the base CPython runtime, so we can use a module variable which is a bit neater.

Ideally we'd eliminate the `Ci_cinderx_initialized` for 3.10 as well but checking a module attribute in the middle of the runtime feels sketchy. I also explored making a global `Ci_cinderx_initialized` variable for use across the CinderX modules in 3.12 but this gets messy. I think it needs to use weak symbols to work, but I couldn't quite get them to do what I wanted. The module variable approach is certainly easy to get to work and understand.

Reviewed By: alexmalyshev

Differential Revision: D59135916

fbshipit-source-id: 528377f0fbc98a92ea2d2180586c57cac70fdaa7