Add definition for CO_STATICALLY_COMPILED and place to put flags like this
Summary:
With CinderX flags like `CO_STATICALLY_COMPILED`  are stored in Python structures but only written or observed in CinderX code. So, it makes sense for their definitions to live in the CInderX codebase.

This diff adds `cinderx/Common/extra-py-flags.h` as a place to put them and initially introduces `CO_STATICALLY_COMPILED` (renamed to `CI_CO_STATICALLY_COMPILED`).

We'll still need to verify the flags are available with each new Python release, but hopefully this is often the case so not much will change. It also keeps them all in one place so it's easy to check everything in one go.

Reviewed By: alexmalyshev

Differential Revision: D59134541

fbshipit-source-id: b96d75a2e8433f7c67922647fa5ad1511e23fe5f