Separate out the core LIR definitions into their own library
Summary:
So that we're not globbing all the LIR files together.  The concepts of Block,
Function, Instruction, and Operand are difficult to tease apart, so they'll
remain together for the near future.

The stuff with lir::Printer doesn't need to be here, and it ultimately pulls in
the HIRPrinter so we should definitely keep them separate.

Reviewed By: DinoV

Differential Revision: D58258091

fbshipit-source-id: 8bbde4bd3f3453fdaab0dcaacb88eddb3b4fb006