Separate Symbolizer out from Runtime
Summary:
The Disassembler and HIRPrinter both expect to be able to call upon the
process-level Symbolizer instance stored in the Runtime.  This then means that
they need to pull in all the dependencies of Runtime... which gets expensive.

This diff adds a new global instance of Symbolizer that has the same lifetime as
Runtime, it gets shutdown in the same place where Runtime::shutdown() is called.
Now Symbolizer and Disassembler can be broken out into their own libraries.

Differential Revision: D58291845

fbshipit-source-id: 9a61f32a15115e46c85c99a96a540c0cb9b98d91