Don't use pycore_atomic.h in CinderX
Summary:
It's an internal header and it conflicts with the standard C++ `atomic` header.
The only place we're using it is in ParallelGC, which isn't upstreamed yet.
When we do try to upstream it, we'll need to switch over to the new `pyatomic.h`
header anyway (not available in the 3.12 branch).

For now use the stdatomic.h header directly in parallel_gc.c.

Reviewed By: jbower-fb

Differential Revision: D59341494

fbshipit-source-id: 2a51caf80013acdc84364f4da9cd952ca31973d3