Use function to get dictionaries from types
Summary:
In 3.12 the place to get dicts for builtin types is the interperter data and functions have been added to abstract this.

This diff mostly dumbly replaces direct reference in CinderX to `PyTypeObject::tp_dict` with `_PyType_GetDict()` (which returns a borrowed type). To keep the code clean I've added a macro to `py-portability.h` that just directly accesses the field for 3.10 use.

Reviewed By: alexmalyshev

Differential Revision: D59143468

fbshipit-source-id: 2e8c2ee3ec4e606b2ac98b591ff8be65df05ea13