Allow native tests to build and _cinderx.so to be loaded
Summary:
The native tests statically link libpython this means all symbols in the library are avaialble. This conflicts with the CinderX modules being dynamically linked on load causing it to fail with missing stubs. So, we need to have stubs for the case where we are dynamically linking but not when statically linking.

Deal with this by making our stubs of functions which *just* need exporting from libpython "weak" so they can be overridden by the real implementations when available.

This is not great because it means the native tests use different code from the other tests. Hopefully this isn't a big deal in practice. The native tests probably aren't exercising these functions much if at all.

Also remove argument names which are suddenly now causing lint errors.

Differential Revision: D59143470

fbshipit-source-id: 976f844e61f9c342426a99502e6d91bc86281f0c