Incorporate dict-common.h into checked_dict, use right header elsewhere
Summary:
For checked dict we don't want to rely upon Python's headers, this is a clean implementation and doesn't need to follow the CPython implementation.

Everywhere else we can use `internal/pycore_dict.h` in place of the old `dict-common.h`.

There's also some small changes to the layout of dict objects which are updated.

Reviewed By: alexmalyshev

Differential Revision: D59136285

fbshipit-source-id: 4faa2cfbb7cb0c50df64322752d8ed085d63f604