Understand binary/inplace ops and exact ints
Summary:
Currently the JIT doesn't infer that an exact int += exact int produces another exact int. This makes it understand that.

We have something similar already in place in simplify.cpp for normal binary ops that will optimize an `exact int + exact int` into producing an `exact int`. But because this runs in simplify in a loop ssa will initially infer an object return type, and then we'll never be able to know that it actually produces an int. So this updates both the in place and bin op cases so that we'll successfully infer the result type to be an int.

Reviewed By: alexmalyshev

Differential Revision: D59343577

fbshipit-source-id: 4072f38bbd78f8a1ec62cbc35bfbe56c5eb1a0c4