Add inplace long optimization that matches bin op optimization
Summary: While working on D59343577 I ran into D33724533 which made me think that this was the correct approach to handling the inplace op case as well. Turns out we still need to know the types in ssa due to loops, but I coded all of this up before discovering it, so may as well improve the code gen for inplace ops.

Reviewed By: alexmalyshev

Differential Revision: D59477911

fbshipit-source-id: 4c610b5139a478b6d2dfe5b65f321f5aa3f2a705