Make RaiseAwaitableError work across 3.10 and 3.12
Summary:
Trying to make HIR construction for GET_AWAITABLE work across both Python
versions.  In 3.10, the error it reports is based on the previous two bytecode
instructions.  In 3.12, the error is based on the oparg of the GET_AWAITABLE.

This diff moves the logic of determining which error to report out of the
runtime helper and up into the HIR builder, which makes the HIR instruction and
associated runtime helper definitions much simpler.

Reviewed By: jbower-fb

Differential Revision: D59472575

fbshipit-source-id: 9dcac823f47a101b5c7d558310786b09b476e7ee