Remove unneeded uses of CHANGED_PYCODEOBJECT asserts
Summary:
The one in deopt is tied to a CHANGED_PYFRAMEOBJECT assert, and that one seems
more relevant given how the code is working with Python frames.

The one in the runtime helpers seems like it's a nop, the zombie frame doesn't
exist in 3.12 and either way it was just doing a prefetch.

Reviewed By: DinoV

Differential Revision: D59492956

fbshipit-source-id: 56e0c1182f122d6120ef1680bffb30dc7a5b1c48