Make first include of Static Python impl files their header
Summary:
C/C++ convention.  Helps ensure that headers can be included anywhere and don't
have transitive dependencies on other headers.

Fixing an issue with _PyDict_NotifyEvent() that shows up for generic_type.c.  It
will get the actual definition of the function in the 3.12 build, which
conflicts with our stub.  This diff makes the stub into a macro that feeds into
the actual definition.

Reviewed By: DinoV

Differential Revision: D59606126

fbshipit-source-id: b652d785bd4009b77e8eacca3d2f4b86e357286e