Enable checked dict on 3.12
Summary:
Follow up from D59136285 - this actually enabled checked dict on 3.12

We bring in the definition for `PyDictObject` as `CiDictObject`. To avoid a massive re-name this just #define's PyDictObject as CiDictObject.

Our implementation is a copy/paste of the 3.10 dictionary implementation. There's some stuff in there that's not used like dictionary watchers where we're calling the API which expects a `PyDictObject` but now we have an incompatible `CiDictObject` so that's basically reverted back to the baseline 3.10 implementation. We could potentially even just remove dictionary versions altogether but I've left it in.

There's also a few checks where we're seeing if we actually have a `PyDict` for things like free lists, and since that's never true it removes those.

`CheckedDict` is also not subclasable so the check for `__missing__` is dropped.

We were also re-using `_PyDict_SizeOf` so that turns into a CinderX-specific version.

Reviewed By: jbower-fb

Differential Revision: D59305184

fbshipit-source-id: 4c6bc4248d12ae5ebefa32a7abf94dd3166ad05c