Intentionally leave HIRBuilder::addInitializeCells() blank for 3.12
Summary:
This is 3.10-only code to initialize cell vars and free vars.  In 3.12 we have
explicit instructions (MAKE_CELL, COPY_FREE_VARS) that do this in the function's
prologue before the RESUME instruction is called.

We'll add support for the new bytecodes once we can wire in the 3.12 opcode
header to the JIT.

Reviewed By: DinoV

Differential Revision: D59694354

fbshipit-source-id: cb9cc1962b937e1af597d8c353e7c064414f2e67