Support using 3.12 bytecodes in CinderX
Summary:
Currently the build always pulls in the 3.10 bytecode definitions, even on the
3.12 build.  This diff adds in the build support for switching to use the 3.12
definitions when using the 3.12 runtime.

The 3.10 bytecodes are stubbed out for the 3.12 build.  This makes it easier to
land the change, as we don't need to add Python version checks throughout the
JIT.  The 3.10 bytecode paths will just silently be unreachable code.  The stubs
use large numbers (40,000+) intentionally so that if we ever do try to use them,
we should hit type errors as they wouldn't fit in a byte.

Differential Revision: D59826975

fbshipit-source-id: 18fa59205803bdd2168b2514d354f1ac72ff35fe