Remove JIT type profiling infra
Summary:
Doesn't feel great removing this but, we're not using it right now and `s_live_types` is causing problems with 3.12 and native tests. The issue there is 3.12 doesn't send type change events when types are deleted so references to types are in `s_live_types` are out living the types causing use-after-free errors when `s_live_types` is deallocated.

I've deleted all the references I could find but I've left some stubs in `cinder.py` from 3.10 just in case. Once we move to 3.12 these will no longer be present and we can deal with any fall out then.

Reviewed By: alexmalyshev

Differential Revision: D59877256

fbshipit-source-id: 7c76c649d8d8dce99eab8bd6744126db708d1a58