Break up "container" from field/method in type descriptor serialization
Summary:
Currently we serialize our type descriptors as just a single long list of where things live, if these were just strings they would look like "module.Class.function" or "module.function" or "module.Class.InnerClass.field" but these are actually stored as tuples so we don't need to split/parse strings (e.g. `("module", "Class", "function")`).

This means when we want to resolve a descriptor we go through this entire path in a rather happenstance  sort of way and deal with whatever we encounter on the way. Instead of doing that this breaks our descriptors into a "container" component as well as a "target" component. The container looks like `("module",)` or `("module", "Class")` or `('package', 'module', 'Class')`. Then the full descriptor looks like `('module', 'Class'), 'function')`.

Ultimately this lets us treat type resolution as a more stand alone thing.

Because this modifies the pyc file format the magic version number is updated.

Reviewed By: alexmalyshev

Differential Revision: D59645035

fbshipit-source-id: 5e9178bb78059e6b7615cf7899f2320d5c08a55c