Make compile errors in RuntimeTests actually debuggable
Summary:
`RuntimeTest::compileAndGet()` and associated functions all propagate up a
boolean status on whether the compilation was successful or not.  When we do hit
a failure, all we have to work with is that the compiled function we tried to
load is null, so it's tricky to know where to begin debugging.

This adds a new `THROW()` macro that will try to print a Python exception if it
exists, and will then also throw a C++ exception with relevant messaging.  So if
there's a Python error we'll see it, and we'll also see in which stage of the
execution process the failure happened.

Reviewed By: DinoV

Differential Revision: D59962992

fbshipit-source-id: 51de9c0619eae6bed014dc2dd673fe106bc21dd2