Move vtable implementation targets
Summary:
We have *a lot* of different v-table entries that get plugged into the v-table when we have things other than simple functions. These handle all of the various delicacies of dispatching to the real function.

This pulls all of those entry points out into a new library that can live on its own. There's one exception: the lazy init entry is what gets populated into the v-table before it's been invoked for the first time. That continues to live in classloader.c.

Reviewed By: alexmalyshev

Differential Revision: D59864306

fbshipit-source-id: 336a0c144180ad0659557662127d32156aff91f3