Move vtable implementation targets
Summary:
Take 2 of D59864306. The previous version of this diff hit some issues with non x64 builds of Cinder and got reverted. The issue is that the v-table thunks (as defined in `vtable_defs.h`) for x64 are just defined as `void thunk_name(void);` and this is fine, these never get called from C code and the prologue is in assembly. There's a backup definition for non-x64 platforms which has argument types though and merely calls the vector call implementation, therefore it needs the args. This aligns the definitions so they all have argument types.

We have a lot of different v-table entries that get plugged into the v-table when we have things other than simple functions. These handle all of the various delicacies of dispatching to the real function.

This pulls all of those entry points out into a new library that can live on its own. There's one exception: the lazy init entry is what gets populated into the v-table before it's been invoked for the first time. That continues to live in classloader.c.

Reviewed By: alexmalyshev

Differential Revision: D60180678

fbshipit-source-id: ca5a1014431607cf9072667d3742869c99eec195