Drop StaticPython's dependency on the entire JIT
Summary:
Take 2 of D59969622, no changes.

The `StaticPython` directory only has one thing it really depends on out of the intertangled ball of code that is the JIT - and that's the global watchers.

This pulls the actual global cache manager into a singleton and moves the associated APIs for accessing it into the `Jit:global-cache`

Reviewed By: alexmalyshev

Differential Revision: D60180675

fbshipit-source-id: d27ba7bc35e0040893150beabac665f0a123619b