Bubble up CinderX initialization failures higher
Summary:
After locally patching in CinderX to initialize for 3.12, I see a lot of:

  RuntimeError: Failed to initialize CinderX

errors in the runtime tests that end in `cinderx_init()`.  This isn't super
helpful, and it turns out this specific error overrides any previous errors that
might have been raised.  Making it fire only when there is no Python exception
already set instead produces:

  SystemError: rand() method: bad call flags

which is still pretty cryptic, but more information that what we started with.

(AFAICT We're failing to create the _static module, that's where the error comes
from)

Differential Revision: D60255454

fbshipit-source-id: 45eb7fae9abc7d7bf1675705d3e0185a667c3cc8