Remove StaticPython-headers
Summary:
The last remaining uses of `StaticPython-headers` are in StaticPython. We can properly depend upon just the things we need and export the things we depend upon.

There's a special snowflake in the form of `xxclasslaoder`. It depends upon generic types and typed defs, and that'll transitively bring in code like our watchers. We don't want `xxclassloader` to refer to this at all, and instead want it to pick up the references from `_cinderx`. So we need to directly bring in the headers that it depends upon to avoid duplicating the compiled code (and static variables) into its .so.

Reviewed By: alexmalyshev

Differential Revision: D59974766

fbshipit-source-id: 91705cca52faa47e5a9334dffab79b0aeb87534c