Apply the same .clang-format file to all of fbcode/cinderx
Summary:
We're currently pulling in the .clang-format file into specific subdirectories.
This meant that we accidentally forgot to symlink it into fbcode/cinderx/Common
and it's never been formatted.

The big directory that was never auto-formatted is fbcode/cinderx/StaticPython.
This makes it consistent with the rest of the codebase.  The overall format
config is optimized for C++ code.  If we want to have StaticPython/ use a format
that's closer to CPython's style, then we should add an
fbcode/cinderx/StaticPython/.clang-format as an override.  Keeping it entirely
unformatted is worse.

Differential Revision: D60173263

fbshipit-source-id: c704b6dc14dacba0e4b4087de9e2cc3a905af9e1