Creating opcode stubs for 3.12 opcodes in 3.10
Summary:
We already had 3.10 opcodes for 3.12 builds in upgrade_stubs.h.  This diff moves
those to a new cinderx/Common/opcode_stubs.h file, and adds 3.12 opcodes over
into 3.10 builds as well.

The idea here is to make the opcode definitions always available across both
versions, to cut down on `#if` checks like:

  #if PY_VERSION_HEX >= 0x030C0000
    if (opcode == RESUME) {
      doSomething();
    }
  #endif

  #if PY_VERSION_HEX < 0x030C0000
    if (opcode == ROT_N) {
      doSomethingElse()
    }
  #endif

across the codebase.  The downside is that this is duplicating what exists in
3.12/Lib/opcode.py and 3.10.cinder/Lib/opcode.py.

To demonstrate this working for 3.10, this diff adds RESUME support to
HIRBuilder, which is compiled across both versions despite the opcode only being
added in 3.11.

Reviewed By: jbower-fb

Differential Revision: D60326775

fbshipit-source-id: cf6fcb4dc6377df176a6e70757c48eb49d814c75