Remove Ci_HookType_add_subclass hook
Summary:
Removing another hook. Instead of hooking into CPython we can track modifications to `tp_subclasses`.

Because the watchers are all listening to all of the dictionaries we want to keep the checks the watcher does to be as cheap as possible. We can quickly filter most modifications by checking the key/value to be a long/weakref pair.

Even if we get those we keep the lookup as cheap as possible by storing in the dictionary a `_Ci_ObjectKey` which will compare true against the dictionary, and we use `_PyDict_GetItem_KnownHash` to pass in the dict's address as the hash as dictionaries are usually not hashable.

On 3.12 we also need to potentially get the subclasses for types shared across interpreters from somewhere else.

Reviewed By: alexmalyshev

Differential Revision: D59893761

fbshipit-source-id: bbd563ab1e3664252d90f269415d8a01b33fae6a