Better JIT check in RuntimeTests
Summary:
The common way that the runtime tests fail when the JIT isn't initialized
properly is that they try to access the code allocator and then crash because
it's a null pointer.  It's a bit of a confusing failure, so this diff moves it
up earlier in the test process and makes it a bit more explicit.

Ideally we'd be using `_PyJIT_IsEnabled()` but that will actually return `false`
even on our working 3.10 JIT tests.  Our JIT configuration needs to be improved
first before we can reliably use that.

Reviewed By: jbower-fb

Differential Revision: D60399332

fbshipit-source-id: 8024590b2334d2e28efe2564e56d60e0a6ecb5e3