Have BytecodeInstructionBlock keep track of the PyCodeObject
Summary:
Currently it only tracks the bytecode buffer i.e. the `_Py_CODEUNIT*`.  I'm
about to make use of some CPython APIs that require the code object itself, so
making this more general.

In some ways the block iterator gets simpler as it no longer needs to increment
both an index and a pointer in lockstep.

Reviewed By: mpage

Differential Revision: D60603158

fbshipit-source-id: 95bd9a1556cda855696d3aab68eddeb3166dcce3