Switch from using _PyObject_GC_Malloc to PyUnstable_Object_GC_NewWithExtraData
Summary: `_PyObject_GC_Malloc` has gone away in 3.12, but I think the reason we're using it - adding extra data to a GC-managed `PyObject*` - is covered by the recently added `PyUnstable_Object_GC_NewWithExtraData`. The future of this [is a little uncertain](https://github.com/python/cpython/issues/103743) but for now it seems to do what we want.

Reviewed By: alexmalyshev

Differential Revision: D60684943

fbshipit-source-id: d6b8fb001502c68805bd7edd31fdec61eb9389bb