Process children in single pass at end of visiting
Summary:
Currently our process of figuring out where variables belong is extremely different than CPython's. The first significant difference is that CPython performs a single pass after visiting the AST.  We traverse each child node after visiting all of its members.

This starts bringing us closer to CPython by having a single pass. We still traverse the nodes in the same order, and this includes some back and forth as we force things to be globals, but it is at least getting closer to CPython.

Reviewed By: alexmalyshev

Differential Revision: D60875403

fbshipit-source-id: ed2d6ee5ed99e3cd28f931c23d4bbb8a5bfde7bf