Simplify free variable handling for comprehension inlining
Summary:
Currently our comprehension processes free variables defined in the comprehension by calling `scope.add_free`. This then calls `scope.add_frees([var])` which does a bunch of logic.

But the actual necessary logic is much simpler - we know the free variable isn't defined in the scope, as we check that, so we must be adding it as a free variable to the scope. So we just do that.

Reviewed By: alexmalyshev

Differential Revision: D60871638

fbshipit-source-id: 80b42606d12eb1816e7dd6ed7f98f7d7736023e6