Stabalize SxS code tests and improve naming
Summary:
When a SxS stdlib test fails it's currently difficult to repro.

If it fails in Sandcastle and you try and re-run the particular set there's a good chance the failing test won't be in the same set because the hash isn't stable. This uses md5 for the hashing based upon the name so the groups will be stable.

When running locally the names for the test functions are ridiculously long and include dots because they end up including a path into the buck-out directory. This makes it impossible to re-run individual tests.

Reviewed By: alexmalyshev

Differential Revision: D61044766

fbshipit-source-id: 09b69fdaff1ab3472e9536e269a21a36a62648f0