Re-enable linting on fbcode/cinderx/PythonLib/__static__
Summary:
Working through PythonLib/ bit by bit as it's complex.

This diff has a number of manual changes, mostly removing unused imports and
adding `noqa: F401` for imports that are meant to be exported further. I also
needed to move some `pyre-ignore`s that got shuffled to the wrong line.

Reviewed By: brittanyrey

Differential Revision: D61233779

fbshipit-source-id: 8bb7000b6f654778dc61594633c6b344058e608b