Support un-instrumenting opcodes
Summary:
This is necessary for jit::BytecodeInstructionBlock to function.  We use
BytecodeInstructionBlock in HIRBuilder where we know instrumented opcodes will
never show up, but we also use it in the preloader.  In the preloader there's no
guarantees about what opcodes are traversed, so it needs to compute the inline
cache sizes correctly.

Reviewed By: jbower-fb

Differential Revision: D61393389

fbshipit-source-id: d3626ada118bb2ee8c8fa3871f90abe1e0da3055