Update _PyInterpreterFrame::prev_instr pointer from JIT code
Summary:
We need this data for reconstructing stack-traces with line-number information. While this is a bit more expensive than having CPython pull this information as-needed, the hope is this will be balanced out by not needing a hook in CPython.

In this implementation we insert an update opcode as the last step in HIR compilation, before opcodes which might run arbitrary further code and thus consume this data.  To keep the cost down, the value is only updated when there is a line-change (not just a bytecode offset change).

A further optimization to reduce cost would be to only update the subset of bytes in the pointer needed. I'll leave that for later though.

Reviewed By: alexmalyshev

Differential Revision: D61253815

fbshipit-source-id: f7d744c9a58d365ed852900f169c019d852bf997