Allow update_hir_expected.py to split expected values into 3.10 and 3.12 #if-def blocks
Summary: Adds some automation for different expected outputs in C++ tests for 3.10 vs. 3.12. Assuming output is usually different for 3.10 and 3.12 this should be okay. If we hit a case where output for 3.10 and 3.12 changes but it's the same for both this will create unnecessary duplication. I imagine this case will be rare.

Reviewed By: alexmalyshev

Differential Revision: D61574289

fbshipit-source-id: 6f9cc708fd5fcf1c8f45ffd721ebb88caddbabbd